§46 ship-correctness-closure: validation sweeps 01/04/05/06 -> 95/0, gcAtEnd 10/10, Tier-B + scanner residuals
22 tasks across 47 source files. Per docs/threads/CVE-AUDIT-RESULTS.md
'Ship-readiness closure' section + SCALEBENCH.md §46.

Validation sweeps r2: 01 (validateDFGClobberize SIGTRAP family) 86/8 ->
95/0; 05 (validateExceptionChecks unchecked-scope sites at VMTraps:686/
ThreadAtomics:1425/ConditionObject:103/ThreadManager:1124/JSObject:4283/
CallData:76) 44/50 -> 95/0; 04+06 -> 95/0. gcAtEnd property-wait-
termination 10/10 (MachineStackMarker SEGV gone).

03 (validateButterflyTagDiscipline) lint WIRED — 80/15 surfaces the §43
Task-8 residual (43 I14 reports: GetByOffset/PutByOffset storage edge from
non-tag-masking producer) as a regression detector. Task-8 DEFERRED:
JSArray exclusion makes the latent concern unreachable; perf upside
~400-500ms recorded.

§45 duplicate-property: misdiagnosis (private @hasOwn rendered without @
in describe()); under-lock getDirectOffset re-probe at reifyStaticProperty
added defensively (useJSThreads-gated).

TSan build: TsanDeferredCtorMember<StructureTransitionTable> forwarder for
tryGetSingleSlotConcurrently.

--cve 60/3/0/2: 3 fails are exit-3 test-side issues (missing useDollarVM
in requireOptions; B1/B2/B10 fixes now spec-correctly throw where tests
expected old behavior). No memory-safety failures; all Tier-A repros pass.

Corpus 95+96/0, identity 40/0, bench all checksums stable ±2% of §45.

+++ Source/JavaScriptCore/bytecode/UnlinkedFunctionExecutable.cpp
@@ -271,8 +271,8 @@ SourceCode UnlinkedFunctionExecutable::linkedSourceCode(const SourceCode& passed
 {
     const SourceCode& parentSource = !m_isBuiltinDefaultClassConstructor ? passedParentSource : BuiltinExecutables::defaultConstructorSourceCode(constructorKind());
     unsigned startColumn = linkedStartColumn(parentSource.startColumn().oneBasedInt());
-    unsigned startOffset = parentSource.startOffset() + m_startOffset;
-    unsigned firstLine = parentSource.firstLine().oneBasedInt() + m_firstLineOffset;
+    unsigned startOffset = static_cast<unsigned>(parentSource.startOffset()) + m_startOffset;
+    unsigned firstLine = static_cast<unsigned>(parentSource.firstLine().oneBasedInt()) + m_firstLineOffset;
     return SourceCode(parentSource.provider(), startOffset, startOffset + m_sourceLength, firstLine, startColumn);
 }
